fix: Calculation 클래스 제거,  Calculator에서 계산 로직 구현 관련이슈: step4-계산기UI
